refactor: test code
